requirements.txt
torch>=2.1.0
transformers>=4.39.3
peft>=0.9.0
bitsandbytes>=0.42.0
accelerate>=0.28.0
datasets>=2.17.0
evaluate>=0.4.0
wandb>=0.16.3
matplotlib>=3.8.0
plotly>=5.19.0
pyyaml>=6.0.1
pandas>=2.2.0
rich>=13.7.0
scikit-learn>=1.3.0
tensorboard>=2.15.0
vllm>=0.3.2
streamlit>=1.31.0
